backend/app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
numpy
pandas
Pillow
PyPDF2
faiss-cpu
openai
PyMuPDF

